projects-new/microservices-demo-app/services/order-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
grpcio==1.62.1
protobuf==4.25.3
pika==1.3.2
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
